requirements.txt
anthropic>=0.18.1
python-dotenv>=1.0.0
pytest>=7.0.0
flake8>=6.0.0
mypy>=1.0.0
black>=23.0.0  # Code formatting
isort>=5.0.0   # Import sorting 